Move unauthenticated page out of App, protect certain routes
