+++ client/src/App.css
@@ -25,10 +25,6 @@ main {
   flex: 1;
 }
 
-.unauthenticated-app {
-  text-align: center;
-}
-
 
 @media only screen and (min-device-width: 667px) {
   header {
