+++ client/src/App.tsx
@@ -1,7 +1,7 @@
 import { Routes, Route } from 'react-router';
-import { AuthProvider, useAuth } from 'hooks';
+import { AuthProvider, ProtectedRoute } from 'hooks';
 import { ProfileBar, AppFooter, AppHeader } from "components";
-import { AuthCallback } from 'views';
+import { AuthCallback, UnauthenticatedApp } from 'views';
 
 import './App.css';
 
@@ -18,9 +18,12 @@ function App() {
 
 function AppContent() {
   return (
-    <main className=" bound-width">
+    <main className="bound-width">
       <Routes>
-        <Route path="/" element={useAuth().getIsSignedIn() ? <AuthenticatedApp /> : <UnauthenticatedApp />} />
+        <Route index element={<UnauthenticatedApp />} />
+        <Route element={<ProtectedRoute />}>
+          <Route path="/home" element={<AuthenticatedApp />} />
+        </Route>
         <Route path="/auth/callback" element={<AuthCallback />} />
       </Routes>
     </main>
@@ -32,14 +35,6 @@ function AuthenticatedApp() {
     <p>
       Welcome known user!
     </p>
-  )
-}
-
-function UnauthenticatedApp() {
-  return (
-    <section className="unauthenticated-app">
-      <p>Please sign in to continue.</p>
-    </section>
   )
 }
