+++ client/src/views/UnauthenticatedApp/UnauthenticatedApp.css
@@ -1,0 +1,3 @@
+.unauthenticated-app {
+  text-align: center;
+}
