+++ client/src/views/UnauthenticatedApp/UnauthenticatedApp.tsx
@@ -1,0 +1,9 @@
+import "./UnauthenticatedApp.css"
+
+export const UnauthenticatedApp = () => {
+  return (
+    <section className="unauthenticated-app">
+      <p>Please sign in to continue.</p>
+    </section>
+  )
+}
