+++ client/src/views/index.js
@@ -1,5 +1,7 @@
 import AuthCallback from "./AuthCallback/AuthCallback";
+import { UnauthenticatedApp } from "./UnauthenticatedApp/UnauthenticatedApp";
 
 export {
-  AuthCallback
+  AuthCallback,
+  UnauthenticatedApp
 }
